backoffice/inicio/views.py
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status
from .models import *
from .serializer import *

# Create your views here.

class TagList(APIView):
    
    def get(self, request):
        tags = Tag.objects.all()
        serializer = TagSerializer(tags, many = True)
        return Response(serializer.data)
        
    def post(self):
        pass
