backoffice/requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.1
Django==2.0.3
djangorestframework==3.7.7
gunicorn==19.6.0
idna==2.6
pew==1.1.2
Pillow==5.0.0
pipenv==10.1.2
psycopg2==2.7.4
pytz==2018.3
requests==2.18.4
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
whitenoise==3.2.1
